Added Header to Dashboard
